Use qualified imports in A-servant-start.
I'm not sure what's going on with (:<|>). If I try to use it unqualified
in the definition of `server`, it suggests that I might want to include
it in either the "import Servant (...)" or "import Servant.Api (...)"
lines, even though it's already there (and even if I put it in the
Servant line as well).