Add logout ability to D.
This also switches to mostly-qualified imports, it's a messy commit.

Bug: you can no longer hit "enter" to login. That's because the button
is no longer a submit button. As a submit button, it refreshed the page,
which isn't really what we'd want I think? But hitting "enter" to login
would be pretty much an essential feature in a real app.

(You can admittedly use "enter" if you tab to the login button.)

Bug: logs don't disappear if you logout. That's because we never reset
model.logs. We could do that on logout, but resetting every piece of
state separately seems like not what we want. I'd suggest a better way
to go might be to put all the login-protected info in a type, and give
model a Maybe one-of-those.